requirements.txt
requests>=2.31.0
PyYAML>=6.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
pydantic>=2.5.0
APScheduler>=3.10.0
pymupdf>=1.23.0
unidecode>=1.3.0
ollama>=0.1.0